Revert "Fix: Slow Walk"
This reverts commit af2e2102861cd6f7f2b08e70216eb5c671325762.

+++ Osiris/Config.h
@@ -111,7 +111,7 @@ public:
         bool LBYBreaker{ false };
         float LBYAngle{ 0.0f };
         bool Slowwalk{ false };
-        float slowwalkspeed{ 0.0f };
+        int slowwalkspeed{ 0 };
         int slowwalkkey{ 0 };
     } antiAim;
